follow material-ui approach, components named appropriately with index.js
